move generated files to a shared crate
