+++ rust/server/src/lib.rs
@@ -1,2 +1,1 @@
-pub mod adl;
 pub mod server;
