+++ rust/server/src/server/routing.rs
@@ -1,12 +1,12 @@
 use poem::endpoint::{DynEndpoint, ToDynEndpoint};
+use poem::session::{CookieConfig, CookieSession};
 use poem::{post, EndpointExt, Route};
 
-use crate::adl::gen::protoapp::apis::ui::ApiRequests;
+use shared::adl::gen::protoapp::apis::ui::ApiRequests;
 
-use super::handlers;
-use super::poem_adl_interop::{new_access_token_checker, RouteExt};
-use super::AppState;
-use poem::session::{CookieConfig, CookieSession};
+use crate::server::handlers;
+use crate::server::poem_adl_interop::{new_access_token_checker, RouteExt};
+use crate::server::AppState;
 
 pub fn build_routes(state: AppState) -> Box<dyn DynEndpoint<Output = poem::Response>> {
     let access_token_checker = new_access_token_checker(state.config.jwt_access_secret.clone());
