+++ rust/shared/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "shared"
+version = "0.1.0"
+edition = "2021"
+description = "ADL generated code for the protoapp application"
+
+[dependencies]
+base64 = { workspace = true }
+rand = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+sea-query = { workspace = true }
+sea-query-binder = { workspace = true }
+sqlx = { workspace = true }
