+++ rust/shared/src/.adl-manifest
@@ -1,0 +1,30 @@
+# manifest @generated by the adl compiler
+adl/gen/adlc/config/mod.rs
+adl/gen/adlc/config/rust.rs
+adl/gen/adlc/mod.rs
+adl/gen/common/db.rs
+adl/gen/common/http.rs
+adl/gen/common/mod.rs
+adl/gen/common/strings.rs
+adl/gen/common/time.rs
+adl/gen/common/ui.rs
+adl/gen/mod.rs
+adl/gen/protoapp/apis/mod.rs
+adl/gen/protoapp/apis/ui.rs
+adl/gen/protoapp/config/mod.rs
+adl/gen/protoapp/config/server.rs
+adl/gen/protoapp/db.rs
+adl/gen/protoapp/mod.rs
+adl/gen/sys/annotations.rs
+adl/gen/sys/mod.rs
+adl/gen/sys/types.rs
+adl/rt/custom/mod.rs
+adl/rt/custom/sys/mod.rs
+adl/rt/custom/sys/types/map.rs
+adl/rt/custom/sys/types/mapentry.rs
+adl/rt/custom/sys/types/maybe.rs
+adl/rt/custom/sys/types/mod.rs
+adl/rt/custom/sys/types/pair.rs
+adl/rt/custom/sys/types/result.rs
+adl/rt/custom/sys/types/set.rs
+adl/rt/mod.rs
