+++ rust/shared/src/adl/custom/common/mod.rs
