+++ rust/shared/src/adl/gen/common/db.rs
