+++ rust/shared/src/adl/gen/common/http.rs
