+++ rust/shared/src/adl/gen/common/mod.rs
