+++ rust/shared/src/adl/gen/common/strings.rs
