+++ rust/shared/src/adl/gen/common/time.rs
