+++ rust/shared/src/adl/gen/common/ui.rs
