+++ rust/shared/src/adl/gen/mod.rs
