+++ rust/shared/src/adl/gen/protoapp/apis/mod.rs
