+++ rust/shared/src/adl/gen/protoapp/apis/ui.rs
