+++ rust/shared/src/adl/gen/protoapp/db.rs
