+++ rust/shared/src/adl/gen/protoapp/mod.rs
