+++ rust/shared/src/adl/rt/custom/sys/types/result.rs
